Add issue templates
A menu of issue categories will be presented when the user begins the issue creation process.

The issue will be prefilled with prompts for the specific information needed for each issue type.

If none of the issue types are applicable, the user can click the "Open a blank issue" link at the bottom of the issue
template chooser page to get the previous issue creation behavior.

The default organization level security policy is linked to for people wanting to report a security vulnerability.

Links are provided to the Arduino Language Reference, Arduino Forum, and Arduino Developers Mailing list to redirect
support requests.

References:
- https://docs.github.com/en/github/building-a-strong-community/about-issue-and-pull-request-templates
- https://docs.github.com/en/github/building-a-strong-community/creating-a-pull-request-template-for-your-repository
- https://docs.github.com/en/github/building-a-strong-community/configuring-issue-templates-for-your-repository#configuring-the-template-chooser