Fix format specifier for printing CAN message ID in CanMsg.h.
Regarding the X format specifier: C standard §7.21.6.1: "The unsigned int
argument is converted to ..." On some platforms (e.g. mbed_nano), uint32_t,
which is what's being printed here, is an unsigned long int, so this code causes
a compiler warning. The fix is to use the format specifiers from §7.8.1,
specifically PRIX32.