license: add missing license headers
Some of the files in the Arduino core were missing the license headers.

They contain code which is either copied from previous Arduino cores
(which were already LGPL-2.1-or-later), or has been submitted to this
repository where most of the code is clearly LGPL-2.1-or-later.  In any
case, all the contributors have either been directly employed by Arduino
or have signed the CLA.

Add the missing LGPL-2.1-or-later headers to the files.