Remove donation links from readme
The Arduino company is no longer soliciting monetary donations. Community members who wish to contribute to the project
still have opportunities to do so via the other options listed under the "Contributions" section of the readme.

Those who wish to make a monetary contribution can still do so by purchasing official Arduino products. For this reason,
that part of the donation request text is preserved. Since the "Donations" heading doesn't make sense after the removal
of the context of other monetary contribution options, the heading is hereby changed to "Support the project".